Make openstackdeployment schema more strict
allow to deploy only services and openstack release we currently
actually support in the operator.
Add more fields to the validation specs with some descriptions and
validations.

Also improve display a bit - show openstackdeployment resources
when listing all resources in a namespace, fix the Age column
and remove columns that are not working currently.

Change-Id: Ib482a37c95c616d5c8fd30c8067245f945601369
Related-Issue: PROD-30885