Image pre-caching
To make an image pre-caching, just after any openstackdeployment
create/update/resume we are starting a cache helper and waiting it to
be completed. If the cache helper is already started we need to
compare a list of images specified in the openstackdeployment with a
list of images from the cache helper. If the lists are different, we
are re-starting the cache helper. A cache helper restart is ok
comparing to an apply because assuming some images those don't need
to be updated are already cached and restart goes quite quickly, so it
is almost no difference with apply but more manageable.  Cache helper
is implemented as a daemonset.

Change-Id: I713c6d2f6236001633b8f664ce727811e09e664c
PROD: https://mirantis.jira.com/browse/PRODX-1325