Do not use kopf sub-handlers
According to the documentation [1] handlers with sub-handlers will be called
several times until all sub-handlers will be finished. Also there is at least
one known issue [2] when sub-handlers are not executed for the latest state
of the watched resource.

To avoid this behaviour this patch removes sub-handlers usage in flavor of
asyncio coroutines. We need to execute coroutines manually if
kopf.HandlerRetryError is raised.

This patch also provides few improvements to avoid unnecessary template render
calls and calling Kube API to get some data.

[1] https://kopf.readthedocs.io/en/latest/handlers/#sub-handlers
[2] https://github.com/zalando-incubator/kopf/issues/279

Change-Id: I03575e91cf725ec695d138ac5c51935c17d2fb78
Closes: PRODX-1899
Closes: PRODX-878