Watch helmbundle create for status
it seems that when helmbundles are processes quickly enough the
on.field watcher may miss created resource completely so the first
time it sees the resource it is already in success=True state and thus
is not reacting to this field at all.

To mitigate this also watch and process status for on.create events
so that we catch those states too.

Related-Issue: PRODX-1899
Change-Id: I4712d066868fadc5c9b5f0adb6f331eca7f431b4