Avoid health status processing without osdpl
Before this patch, a check if openstackdeployment exists was
in report_to_osdpl. That is kind of a last step and makes no sense to
do any actions if no osdpl. Let's move osdpl in an event handler
itself. Osdpl is also required to make nova cell setup in a next patch.

Change-Id: Idae30ba7144be5a2304e4e970004df0ff8d4ea26
PROD: https://mirantis.jira.com/browse/PRODX-1189