Check child status in current object instead of body
In order to avoid races we need to check status in current object.
For example in case of upgrade from rocky to stein, when placement
is added as separate service, its status can be set to Unknown in
the end of upgrade process because before upgrade body didn't
contain placement field in status.

Related-Prod: https://mirantis.jira.com/browse/PRODX-3218
Change-Id: I3a3ec1fdedee438dc425538c369c0b1ee3b01c42