Fix openstack_version validation in admission
due to programming error the OsDplValidationFailed exception was not
handled properly for the "openstack" validator that validates the
OpenStack version and its transitions.

As a result, the validate call returned HTTP 500 instead of 200 +
allowed=False in the body, and was ignored by K8s, passing thru
inappropriate OpenStack versions on create or update.

Closes-Issue: https://mirantis.jira.com/browse/PRODX-6347
Change-Id: Ia2ec55e194ddb2f032c86783f8bd293268f84808