Use single handle function for OsDpl
ensures that only one handler updates the status of OsDpl

Also, use native Kopf facilities to patch the resource being handled.
This way there'd be a single patch call made by kopf itself (when it updates
its annotation and status) that will also update whatever we put in
"patch" that is passed by kopf to the handler.

Related-Issue: PRODX-7472
Change-Id: Id1efbb5b9ea12f35633c784246ff4e6b947b40e6