Add hash_fields to child objects
The hash_fields allows to identify release values that will lead to
object change in jsonpath-ng parse format. The list of fileds will
be extended in future for all kind of immutable objects.

With this patch we always trying to update ceph metadata from shared
namespace with osdpl changed on on controller reboot.

Related-Prod: PRODX-10819
Change-Id: Ic81c1514782890ab7eb756b26dcdcd1732f54d37