Add known node roles
this patch adds a new enum for 'known' top-level roles a Kubernetes node
can have in relation to OpenStack running on it.

Based on that, the format of OSCTL_OPENSTACK_NODE_LABELS was changed
to be able to specify which labels are expected from certain role.

The node object was made smarter by adding a method with which to query
it if it has a certain role.

This new functionality now used in node and node maintenance request
controllers.

Related-Issue: PRODX-11825
Change-Id: Idffaacb95926f9a896f9a4d2d5eb07f9230e5bce