Add spec:features:nova:vcpu_type setting
this is a single option that will translate to a proper combination of
libvirt.cpu_mode and libvirt.cpu_model(s) depending in value and
openstack version:

- for host-model and host-passthrough - set cpu_mode to value
- for others - set cpu_mode: custom and cpu_models to value
  - for openstack <= stein - set cpu_model option instead

The admission controller was also amended to not allow specifying a
coma-separated list of values when openstack is Stein or older, or when
either host-model or host-passthrough is provided.

Default is changed to "host-model" to provide the optimal compromize
between performance / CPU features available to VM and live-migration
stability.

Related-Issue: PRODX-18910
Change-Id: I2196d41da916959df78944a4a35876ebd27b25ae