Neutron: enable live_migration_events
When reverting a resized instance, Nova will timeout
waiting for the vif-plugged external event. The event
is sent as soon as Nova updates the port to the original
source. This happens before Nova gets ready to wait for
events, so the event arrives too soon and Nova times out.

Relared-PRODX: PRODX-18760
Change-Id: I6df73af156e0f3cf97ef54fb1ad28dc9d2d00b1a