Do not send audit events to panko
these are generated on every access to any API by the
keystonemiddleware, but IMO are not very useful in _tenant-facing_
telemetry. At the same time, they constitue about 80% of all events
(measured as a result of full tempest run on the environment), and slow
panko down considerably.

Let's not keep them in panko.

Related-Issue: PRODX-19007
Change-Id: I82aefe08e484e3b493acb0ec85e1c5c258018332