Set trusted dashboard as multistring option
in order to allow customers to have more than one trusted dashboard
specified, we need to declare trusted_dashboard option as a multistirng
one in the identity service template,
otherwise even specyfying it in the 'spec:services' won't work due to
us not overriding dict values with different type (dict over string in
this case).

With this patch, specifying of multiple trusted dashboards in
'spec:services' will work, however ALL values for trusted dashboards,
including one that would have been auto-generated by the template,
must be explicitly set in the 'spec:services' as we do not merge lists.

Change-Id: Id8b28556699f6a3cbfb893ea86c22de0009cb7fb