Harmonize keystone fernet token defaults
currently our defaults are not self-consistent for using service
auth tokens, which we explicitly enable for Cinder and Nova,
and which are always in use by e.g. Ironic.

In order to make our default settings consistent,
we need to set max_active_keys to 7.
This patch also copies defaults for other related options
from keystone chart to controller to have them explicitly there.

Related-Issue: PRODX-21376
Change-Id: I2ee042809d2ee2fdc380ca8edc795a4fc76db6a9