Set versioned identity endpoint in Horizon templates
apparently any federated auth plugin does not work when
the identity endpoint is unversioned.

Since this affects client libraries where we have no control over
what users of OpenStack use, we can not effectively fix this in affected
libraries (keystoneauth) in downstream only, as users will still
install those from PyPI, not our downstream sources.

Thus as a workaround, we set those endpoints to versioned ones for the
case of federated identity in the templates for openrc and clouds.yaml
files when they are generated by Horizon.

Upstream bug to follow:
https://bugs.launchpad.net/keystoneauth/+bug/1998366

Related-Issue: PRODX-28577
Change-Id: I30cce179654ea0e9377ad261ebee7ceb79b7f042