Flip the default of OIDCClaimDelimiter
keystone expects that list values are always delimited by semicolon,
however the default in mod-auth-openidc is a coma.
This discrepancy forces users to write overly complex regexes in their
mappings to split the string to array, which is obviously suboptimal
as it is error-prone as any regex.

After two releases of warnings and deprecations we finally flip
the default of this option to what Keystone expects.

Closes-Issue: PRODX-25957
Change-Id: I6c6ce48ce826faa00b0b314b4c71775fc7aab87d