Remove filter-validation if plugin VPNDriverPlugin enabled
Due to the fact that service plugin VPNDriverPlugin doesn't support filter-validation
plugin it is disabled automatically in neutron. However tempest config still
has filter-validation in neutron extension list and tests relay to it fails.
Therefore to allign this filter-validation is removed from extension list
if VPNDriverPlugin is enabled.

Related-PRODX: PRODX-29354
Change-Id: Iffa924a3d2c92262b15249b5f72ef81aa164e413