Use service role for tempest admins with strict_admin
the strict_admin feature restricts 'admin' role in 'admin' project
to be the admin, but when tempest creates its 'admin' users it does so
in 'tempest' domain, so they are not fit for various elevated calls.

Instead, use 'service' role that for backward compatibility reasons
is still enough to do anything regardless of the scope.

Closes-Issue: PRODX-34431
Change-Id: I755475aff3ae830d46681e8adfb77d65a011f5d6