Implement nodedisable fuctionality
Operator may pause LCM on the specific node, when it is
done NodeDisableNotification object is created by LCM.
In this state we do not care about the node liveness during
LCM operations. When node is disabled LCM removes it from
swarm and kubernetes nodes, this patch updates handling
k8s node removal, the metadata (nodeworkloadlock) is removed
only when nodedisablenotification is not present. Otherwise
postpone metadata removal on nodedisablenotification removal.

Related-Prod: PRODX-37988
Change-Id: Ia9a5db454015f624d400d383a4c1b106ac371c02