Keystone's memcache: flush on reconnect
It prevents a scenario that can cause stale data
to be read. If there's more than one memcached server
and the connection to one is interrupted, keys that
mapped to that server will get reassigned to another.
If the first server comes back, those keys will map
to it again. Get can read stale data.

Related-PRODX: PRODX-38436
Change-Id: I1c978a959731408eec3172b09c69361ec2155404