Use 2 glance-api workers in tiny size
in Glance, we have a custom process-based attachment manager, which is
auto-enabled when there's more than one glance-api workers.
Since our CI is using the `tiny` size of OsDpl, effectively currently
this attachment manager is not being tested despite being used in
real deployments.

Previously we had problems with running 4 api workers in CI
(which is why it was scaled down to 1), let's try 2 instead.

Related-Issue: PRODX-38792
Change-Id: I14405204da97c9bc5c6192d1fd6cf5cde621bcdd