Skip one new test when sig verification is on
`test_minimum_basic_instance_hard_reboot_after_vol_snap_deletion`
is a new test in Antelope, next to `test_minimum_basic_scenario`
that we already skip when image signature verification is enabled
in Nova.
These tests can not work in such environmants since they both
create the image in Glance as part of the test, and do not sign it.

Closes-Issue: PRODX-37477
Change-Id: I3944a69a6a9c0e65eb04359f2e20ca6bf464896a