Handle removed node in safe maner
Recently NodeDisableNotification was introduced, which is a way
to mark node as non operational. During this notification the k8s
node object disappear as soon user mark it. This patch makes
code that works with k8s nodes more safe. Also keep original
node object in nodeworkloadlock, as it is the only one way
to identify role of the node.

Related-Prod: PRODX-40367
Change-Id: I62c35524928a18067d6a937d03bccd4a9bb63d18