[functional] Get console output on server delete
We do not know why metrics are not changed, may be host is
not booted in time yet. Log console output for server to
be able to identify if server is actually booted.

Increase success metric waiter for 30 seconds, it is time
that ubuntu takes to start network.

Related-Prod: PRODX-41176
Change-Id: I622adadd0423eb938b1593b8ee2fe2ad44816b46