Add osdplst lcm progress field
The lcm progress shows the current status of applications that
are deployed with osdpl. An Application is a set of k8s objects
the aplication is ready when all k8s objects are applied to
latest version. The LCP progress is x/y where x is number of
ready applications and y is total number of applications.

Related-Prod: PRODX-3456
Change-Id: I83f1bfde5124682be190371f10a7ce7cdb06966a