Drop domain and user domain from default idp mapping
the user domain was actually ignored until Caracal,
and users were created in the domain of IdP anyway.
However, it started to actually work in Caracal,
so w/o this change, the new incoming users will now
be created in Default domain.. which is not good.

The domain one is basically noop, as it defines the domain
for possible user, group or project stanzas,
so if we keep it in place, it will potentially still override
the user domain.

Related-Issue: PRODX-43737
Change-Id: I8908922b73801f8216ca8e1849b59b568b996ff6